pyproject.toml
[tool.poetry]
name = "krules-dev-support"
version = "0.12.17"
description = ""
authors = ["Airspot <info@airspot.tech>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "krules_dev"}]

[tool.poetry.dependencies]
python = ">=3.10"
jinja2 = ">=3.0.1"
sane-build = ">=7.2"
sh = ">=2.0.4"
structlog = ">=23.1.0"
rich = ">=13.5.2"
better-exceptions = ">=0.3.3"
google-auth = ">=2.22.0"
google-cloud-secret-manager = ">=2.16.4"
python-dotenv = ">=0.19.0"
pyyaml = ">=6.0.1"
pulumi = ">=3.0.0,<4.0.0"
pulumi-gcp = ">=6.33.0"
pulumi_docker = ">=4.4.1"
pulumi-kubernetes = ">=4.3.0"
pulumi-google-native = ">=0.31.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
